UI refactoring
all UI components and controllers have been moved into src/ui/
