+++ src/ui/new_project_widget.cpp
@@ -1,0 +1,137 @@
+#include "new_project_widget.h"
+
+NewProjectWidget::NewProjectWidget(MainWindow* parent, OptionManager* options)
+        : main_window_(parent), options_(options), prev_selected_(false) {
+    setWindowFlags(Qt::Dialog);
+    setWindowModality(Qt::ApplicationModal);
+    setWindowTitle("New project");
+
+    project_name_text_ = new QLineEdit(this);
+    project_name_text_->setText(QString::fromStdString(*options_->project_name));
+
+    QPushButton* project_path_select = new QPushButton(tr("Select"), this);
+    connect(project_path_select, &QPushButton::released, this,
+            &NewProjectWidget::SelectProjectPath);
+    project_path_text_ = new QLineEdit(this);
+    project_path_text_->setText(QString::fromStdString(*options_->project_path));
+
+    QPushButton* image_path_select = new QPushButton(tr("Select"), this);
+    connect(image_path_select, &QPushButton::released, this,
+            &NewProjectWidget::SelectImagePath);
+    image_path_text_ = new QLineEdit(this);
+    image_path_text_->setText(QString::fromStdString(*options_->image_path));
+
+    QPushButton* create_button = new QPushButton(tr("Create"), this);
+    connect(create_button, &QPushButton::released, this,
+            &NewProjectWidget::Create);
+
+    QGridLayout* grid = new QGridLayout(this);
+
+    grid->addWidget(new QLabel(tr("Project name"), this), 0, 0);
+    grid->addWidget(project_name_text_, 0, 1);
+
+    grid->addWidget(new QLabel(tr("Project folder"), this), 1, 0);
+    grid->addWidget(project_path_text_, 1, 1);
+    grid->addWidget(project_path_select, 1, 3);
+
+    grid->addWidget(new QLabel(tr("Images folder"), this), 2, 0);
+    grid->addWidget(image_path_text_, 2, 1);
+    grid->addWidget(image_path_select, 2, 3);
+
+    grid->addWidget(create_button, 3, 3);
+}
+
+bool NewProjectWidget::PathsValid() {
+    return boost::filesystem::is_directory(ImagePath()) &&
+           boost::filesystem::is_directory(ProjectParentPath());
+}
+
+std::string NewProjectWidget::ProjectName() const {
+    return project_name_text_->text().toUtf8().constData();
+}
+
+std::string NewProjectWidget::ProjectParentPath() const {
+    return EnsureTrailingSlash(project_path_text_->text().toUtf8().constData());
+}
+
+std::string NewProjectWidget::ProjectPath() const {
+    return  ProjectParentPath() + EnsureTrailingSlash(ProjectName());
+}
+
+std::string NewProjectWidget::ImagePath() const {
+    return EnsureTrailingSlash(image_path_text_->text().toUtf8().constData());
+}
+
+void NewProjectWidget::SetProjectPath(const std::string& path) {
+    project_path_text_->setText(QString::fromStdString(path));
+}
+
+void NewProjectWidget::SetImagePath(const std::string& path) {
+    image_path_text_->setText(QString::fromStdString(path));
+}
+
+void NewProjectWidget::Create() {
+    if (ProjectName() == "") {
+        QMessageBox::critical(this, "", tr("You must set project name!"));
+        return;
+    }
+
+    if (!PathsValid()) {
+        QMessageBox::critical(this, "", tr("Invalid paths."));
+        return;
+    }
+
+    boost::filesystem::path path(ProjectPath());
+
+    if (boost::filesystem::is_directory(path)) {
+        QMessageBox::critical(this, "", tr("Project name corresponds with existent directory.\nPlease choose another project name / project directory"));
+        return;
+    }
+
+    boost::filesystem::create_directory(path);
+
+    if (main_window_->mapper_controller->NumModels() > 0) {
+        if (!main_window_->OverwriteReconstruction()) {
+            return;
+        }
+    }
+
+    *options_->project_name = ProjectName();
+    *options_->project_path = ProjectPath();
+    *options_->database_path = ProjectPath() + "base.db";
+    *options_->image_path = ImagePath();
+
+    Database database;
+    database.Open(*options_->database_path);
+
+    main_window_->WriteProjectConfiguration();
+
+    main_window_->UpdateProjectInfoStatusBar();
+
+    hide();
+}
+
+void NewProjectWidget::SelectProjectPath() {
+    project_path_text_->setText(QFileDialog::getExistingDirectory(
+            this, tr("Select a directory for project..."), DefaultDirectory(),
+            QFileDialog::ShowDirsOnly));
+}
+
+void NewProjectWidget::SelectImagePath() {
+    image_path_text_->setText(QFileDialog::getExistingDirectory(
+            this, tr("Select a directory with images..."), DefaultDirectory(),
+            QFileDialog::ShowDirsOnly));
+}
+
+QString NewProjectWidget::DefaultDirectory() {
+    std::string directory_path = "";
+    if (!prev_selected_ && !options_->project_path->empty()) {
+        const boost::filesystem::path parent_path =
+                boost::filesystem::path(*options_->project_path).parent_path();
+        if (boost::filesystem::is_directory(parent_path)) {
+            directory_path = parent_path.string();
+        }
+    }
+    prev_selected_ = true;
+    return QString::fromStdString(directory_path);
+}
